Explicitly truncate status effect sub power

+++ src/map/lua/lua_baseentity.cpp
@@ -13298,7 +13298,7 @@ bool CLuaBaseEntity::addStatusEffect(sol::variadic_args va)
 
         // Optional
         auto subType         = va[4].is<uint32>() ? va[4].as<uint32>() : 0;
-        auto subPower        = va[5].is<uint16>() ? va[5].as<uint16>() : 0;
+        auto subPower        = va[5].is<double>() ? static_cast<uint16>(va[5].as<double>()) : 0;
         auto tier            = va[6].is<uint16>() ? va[6].as<uint16>() : 0;
         auto sourceType      = va[7].is<EffectSourceType>() ? va[7].as<EffectSourceType>() : EffectSourceType::SOURCE_NONE;
         auto sourceTypeParam = va[8].is<uint16>() ? va[8].as<uint16>() : 0;
@@ -13365,7 +13365,7 @@ bool CLuaBaseEntity::addStatusEffectEx(sol::variadic_args va)
 
     // Optional
     auto subType    = va[5].is<uint32>() ? va[5].as<uint32>() : 0;
-    auto subPower   = va[6].is<uint16>() ? va[6].as<uint16>() : 0;
+    auto subPower   = va[6].is<double>() ? static_cast<uint16>(va[6].as<double>()) : 0;
     auto tier       = va[7].is<uint16>() ? va[7].as<uint16>() : 0;
     auto effectFlag = va[8].is<uint32>() ? va[8].as<uint32>() : 0;
